Use a ConfigMap to generate vanity url.
This commit adds support for using a Kubernetes ConfigMap to populate
server_name aliases for a given ingress Name and Namespace. The config map name
is hard coded to `FarvaVanityUrls`. The expected format is:

    "app.vanity.com": "ing1.ns1"

Where `app.vanity.com` is the vanity name we want to expose and "ing1.ns1"
corresponds to a name of "ing1" and namespace of "ns1".

Multiple vanity urls can be created for a ingress single name and namespace.